+++ beatmap/v2/info.ts
@@ -57,51 +57,67 @@ export class Info extends WrapInfo<IInfo> {
       this.environmentNames = data._environmentNames?.map((e) => e) ?? [];
       this.colorSchemes = data._colorSchemes?.map((e) => {
          return {
-            useOverride: e.useOverride ?? false,
+            useOverride: e.useOverride || false,
             colorScheme: {
-               name: e.colorScheme.colorSchemeId ?? '',
+               name: e.colorScheme?.colorSchemeId || '',
                saberLeftColor: {
-                  r: e.colorScheme.saberAColor.r ?? 0,
-                  g: e.colorScheme.saberAColor.g ?? 0,
-                  b: e.colorScheme.saberAColor.b ?? 0,
-                  a: e.colorScheme.saberAColor.a ?? 1,
+                  r: e.colorScheme?.saberAColor?.r ?? 0,
+                  g: e.colorScheme?.saberAColor?.g ?? 0,
+                  b: e.colorScheme?.saberAColor?.b ?? 0,
+                  a: e.colorScheme?.saberAColor?.a ?? 1,
                },
                saberRightColor: {
-                  r: e.colorScheme.saberBColor.r ?? 0,
-                  g: e.colorScheme.saberBColor.g ?? 0,
-                  b: e.colorScheme.saberBColor.b ?? 0,
-                  a: e.colorScheme.saberBColor.a ?? 1,
+                  r: e.colorScheme?.saberBColor?.r ?? 0,
+                  g: e.colorScheme?.saberBColor?.g ?? 0,
+                  b: e.colorScheme?.saberBColor?.b ?? 0,
+                  a: e.colorScheme?.saberBColor?.a ?? 1,
                },
                environment0Color: {
-                  r: e.colorScheme.environmentColor0.r ?? 0,
-                  g: e.colorScheme.environmentColor0.g ?? 0,
-                  b: e.colorScheme.environmentColor0.b ?? 0,
-                  a: e.colorScheme.environmentColor0.a ?? 1,
+                  r: e.colorScheme?.environmentColor0?.r ?? 0,
+                  g: e.colorScheme?.environmentColor0?.g ?? 0,
+                  b: e.colorScheme?.environmentColor0?.b ?? 0,
+                  a: e.colorScheme?.environmentColor0?.a ?? 1,
                },
                environment1Color: {
-                  r: e.colorScheme.environmentColor1.r ?? 0,
-                  g: e.colorScheme.environmentColor1.g ?? 0,
-                  b: e.colorScheme.environmentColor1.b ?? 0,
-                  a: e.colorScheme.environmentColor1.a ?? 1,
+                  r: e.colorScheme?.environmentColor1?.r ?? 0,
+                  g: e.colorScheme?.environmentColor1?.g ?? 0,
+                  b: e.colorScheme?.environmentColor1?.b ?? 0,
+                  a: e.colorScheme?.environmentColor1?.a ?? 1,
                },
+               environmentWColor: e.colorScheme?.environmentColorW
+                  ? {
+                     r: e.colorScheme?.environmentColorW?.r ?? 0,
+                     g: e.colorScheme?.environmentColorW?.g ?? 0,
+                     b: e.colorScheme?.environmentColorW?.b ?? 0,
+                     a: e.colorScheme?.environmentColorW?.a ?? 1,
+                  }
+                  : undefined,
                obstaclesColor: {
-                  r: e.colorScheme.obstaclesColor.r ?? 0,
-                  g: e.colorScheme.obstaclesColor.g ?? 0,
-                  b: e.colorScheme.obstaclesColor.b ?? 0,
-                  a: e.colorScheme.obstaclesColor.a ?? 1,
+                  r: e.colorScheme?.obstaclesColor?.r ?? 0,
+                  g: e.colorScheme?.obstaclesColor?.g ?? 0,
+                  b: e.colorScheme?.obstaclesColor?.b ?? 0,
+                  a: e.colorScheme?.obstaclesColor?.a ?? 1,
                },
                environment0ColorBoost: {
-                  r: e.colorScheme.environmentColor0Boost.r ?? 0,
-                  g: e.colorScheme.environmentColor0Boost.g ?? 0,
-                  b: e.colorScheme.environmentColor0Boost.b ?? 0,
-                  a: e.colorScheme.environmentColor0Boost.a ?? 1,
+                  r: e.colorScheme?.environmentColor0Boost?.r ?? 0,
+                  g: e.colorScheme?.environmentColor0Boost?.g ?? 0,
+                  b: e.colorScheme?.environmentColor0Boost?.b ?? 0,
+                  a: e.colorScheme?.environmentColor0Boost?.a ?? 1,
                },
                environment1ColorBoost: {
-                  r: e.colorScheme.environmentColor1Boost.r ?? 0,
-                  g: e.colorScheme.environmentColor1Boost.g ?? 0,
-                  b: e.colorScheme.environmentColor1Boost.b ?? 0,
-                  a: e.colorScheme.environmentColor1Boost.a ?? 1,
+                  r: e.colorScheme?.environmentColor1Boost?.r ?? 0,
+                  g: e.colorScheme?.environmentColor1Boost?.g ?? 0,
+                  b: e.colorScheme?.environmentColor1Boost?.b ?? 0,
+                  a: e.colorScheme?.environmentColor1Boost?.a ?? 1,
                },
+               environmentWColorBoost: e.colorScheme?.environmentColorWBoost
+                  ? {
+                     r: e.colorScheme?.environmentColorWBoost?.r ?? 0,
+                     g: e.colorScheme?.environmentColorWBoost?.g ?? 0,
+                     b: e.colorScheme?.environmentColorWBoost?.b ?? 0,
+                     a: e.colorScheme?.environmentColorWBoost?.a ?? 1,
+                  }
+                  : undefined,
             },
          };
       }) ?? [];
@@ -137,7 +153,7 @@ export class Info extends WrapInfo<IInfo> {
          _allDirectionsEnvironmentName: this.allDirectionsEnvironmentName,
          _environmentNames: this.environmentNames.map((e) => e),
          _colorSchemes: this.colorSchemes.map((e) => {
-            return {
+            const cs: IInfo['_colorSchemes'][number] = {
                useOverride: e.useOverride,
                colorScheme: {
                   colorSchemeId: e.colorScheme.name,
@@ -150,6 +166,15 @@ export class Info extends WrapInfo<IInfo> {
                   environmentColor1Boost: deepCopy(e.colorScheme.environment1ColorBoost),
                },
             };
+            if (e.colorScheme.environmentWColor) {
+               cs.colorScheme.environmentColorW = deepCopy(e.colorScheme.environmentWColor);
+            }
+            if (e.colorScheme.environmentWColorBoost) {
+               cs.colorScheme.environmentColorWBoost = deepCopy(
+                  e.colorScheme.environmentWColorBoost,
+               );
+            }
+            return cs;
          }),
          _customData: this.customData,
          _difficultyBeatmapSets: Object.entries(